feat(indexing): Cleanup
